reorganize folders of admin components tests
