+++ app/src/test/admin/components/ModellingGroups/SingleGroup/Details/GroupMembersSummaryTests.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import {expect} from "chai";
-import {mockModellingGroupDetails, mockUser} from "../../../../mocks/mockModels";
+import {mockModellingGroupDetails, mockUser} from "../../../../../mocks/mockModels";
 import {shallow} from "enzyme";
-import {GroupMembersSummary} from "../../../../../main/admin/components/ModellingGroups/SingleGroup/Details/GroupMembersSummary";
-import {InternalLink} from "../../../../../main/shared/components/InternalLink";
+import {GroupMembersSummary} from "../../../../../../main/admin/components/ModellingGroups/SingleGroup/Details/GroupMembersSummary";
+import {InternalLink} from "../../../../../../main/shared/components/InternalLink";
 
 describe("GroupMembersSummary", () => {
     it("renders no members if group has no members", () => {
