+++ app/src/test/admin/components/ModellingGroups/SingleGroup/Details/ModellingGroupDetailsPageTests.tsx
@@ -1,0 +1,8 @@
+import * as React from "react";
+import { expect } from "chai";
+import {shallow} from "enzyme";
+
+import "../../../../../helper";
+import { Sandbox } from "../../../../../Sandbox";
+
+
